Add helpful output on create
Since the first thing one would desire after creating a new
Git repository would be to clone it, we now provide a nifty
output that includes the read/write URL of the remote repo.
The output can easily be copy/pasted into a prompt to
initiate a clone of the repo and get started writing code faster.